Add ability to send RE state updates to websocket clients
This functionality is required in order to allow clients
such as APIs and their clients, like GUIs, to receive live
messages whenever the state of the Bluesky RunEngine changes
without requiring them to constantly poll this service.

Note: a better implementation would involve the use of the
queue library so that the below race condition would not
matter. Additionally it would be ideal if the payload that
the websocket servers send their subscribers was formulated
from information that came from the state hook function and
not from them interrogating the run engines state at the
time.

There is a known race condition that manifests whenever the
state of the run engine changes in quick succession however
this is not expected to be an issue in real world scenarios
considering that the purpose is to display upto date info
to the user and if the state changes from A to B then B back
to A in the space of say 2ms the user would not notice even if
the UI updated that quickly.
For a detailed flowchart of the issue see the accompanying
[png image](/readme-git-assets/known_race_condition_when_scan_too_fast.png)