Check supplied function signature looks vaguely ok
We require that any supplied scan (bluesky plan)
function has the run engine as its first parameter
and a state hook function as its second parameter
followed by optional named arguments which are the
parameters passed when invoking the scan function.

This change will cause the dispatcher to exit
immediately if a bad scan function is supplied to
it but the inspection of the function is only on
the functions parameters, it does not perform any
sanity checking on the function body's logic.