Set minimum supported python version to 3.6
because we made earlier specific changes so as to be
compatible with python 3.6 due to constraints imposed
by our deployment environment. This change downgrades
the original minimum version specified in the poetry
build config of 3.7 down to 3.6 because I suspect
there might be further similar future constraints and
the code already works with 3.6 anyway.